Home_Work_12/pyproject.toml
[tool.poetry]
name = "home-work-12"
version = "0.1.0"
description = ""
authors = ["andriymyr <andriymyr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
sqlalchemy = "^2.0.25"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-dotenv = "^1.0.1"
phonenumbers = "^8.13.28"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-extra-types = {version = "^2.4.1", extras = ["phone_number"]}
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
